feat(tests): reorganize init library
