+++ init/cleanup.js
@@ -1,20 +1,4 @@
-const fs = require("fs");
-const Strapi = require("@strapi/strapi");
-
-const cleanUpStrapi = async () => {
-  await Strapi().load();
-
-  const dbSettings = strapi.config.get("database.connection");
-
-  await strapi.destroy();
-
-  if (dbSettings && dbSettings.connection && dbSettings.connection.filename) {
-    const tmpDbFile = dbSettings.connection.filename;
-    if (fs.existsSync(tmpDbFile)) {
-      fs.unlinkSync(tmpDbFile);
-    }
-  }
-};
+const { cleanUpStrapi }  = require("./common");
 
 const run = async () => {
   await cleanUpStrapi();
