+++ init/set-permissions.js
@@ -1,31 +1,6 @@
-const config = require("./config");
-const Strapi = require("@strapi/strapi");
-
 const authenticatedRolePermissions = require("./permissions/authenticated");
 const publicRolePermissions = require("./permissions/public");
-
-const setupStrapi = async () => {
-  await Strapi().load();
-  instance = strapi;
-  return instance;
-};
-
-const stopStrapi = async (instance) => {
-  await instance.stop();
-};
-
-const grantPrivileges = async (name, permissions) => {
-  const role = await strapi
-    .query("plugin::users-permissions.role")
-    .findOne({ where: { name } }, []);
-
-  const roleId = role.id;
-  await strapi
-    .plugin("users-permissions")
-    .service("role")
-    .updateRole(roleId, permissions);
-  return roleId;
-};
+const { setupStrapi, grantPrivileges, stopStrapi } = require("./common");
 
 const run = async () => {
   const instance = await setupStrapi();
