+++ src/api/comment/content-types/comment/lifecycles.js
@@ -9,7 +9,7 @@ module.exports = {
     
     await strapi.entityService.create('api::notification.notification', {
       data: {
-        Title: commentAuthor + " commented on " + comment?.ideaName,
+        Title: commentAuthor + " commented on " + comment?.idea_card?.ideaName,
         Content: commentText,
         entityType: "Comment",
         entityId: commentId,
